feat: rename file weatherAPI/index
